// modules/logistica/estaciones/types/inventarios.ts
export type InventarioEstaciones = {
    estacion: string
    inventarioMagna: number
    pVentasDiarias: number
    dInventariados1?: number
    inventarioPremium: number
    pVentasDiarias2: number
    dInventariados2?: number
    inventarioDiesel: number
    pVentasDiarias3: number
    dInventariados3?: number
}